requirements.txt
# --- Core Engine ---
google-genai==0.3.0
python-dotenv==1.0.1

# --- Security Scanners (Agent D/E) ---
bandit==1.7.10
semgrep==1.93.0

# --- Utilities ---
pytest==8.3.3